api/requirements.txt
faiss_cpu==1.8.0
Flask==2.2.2
numpy~=2.0.0
pandas==2.0.3
gunicorn==22.0.0
Werkzeug==3.0.3
